% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vidProcessing.R
\name{fitModelKNN_CV}
\alias{fitModelKNN_CV}
\title{Generate KNN classifier with LOOCV}
\usage{
fitModelKNN_CV(trajDataFrame, labelVector, kVal)
}
\arguments{
\item{trajDataFrame}{A data frame with measures for each trajectory. This can be produced by
function trajMeasures}

\item{labelVector}{A vector of labels for each trajectory. Its length must be equal
to the number of rows in trajDataFrame}

\item{kVal}{The number of neighbors used for classification}
}
\value{
A vector of classified labels for each trajectory
}
\description{
Fits trajectory data into a K-nearest neighbors classifier using
leave one out cross validation.
}
\examples{
data <- as.data.frame(matrix(1:4,nrow = 2,ncol = 2))
labels <- c("live","dead")
classifierKNN_CV<- fitModelKNN_CV(data,labels,3)
}
\author{
Ayan Bandyopadhyay, Bellarmine College Prep 11/26/2015
}

