% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vidProcessing.R
\name{generateTraj}
\alias{generateTraj}
\title{Generate a \code{TrajectorySet} object with corrected y-values}
\usage{
generateTraj(particles, L = 26, R = 3, epsilon1 = 0, epsilon2 = 0,
  lambda1 = 1, lambda2 = 0, penaltyFunction = penaltyFunctionGenerator(),
  include.area = FALSE, frames)
}
\arguments{
\item{particles}{A \code{ParticleSet} object}

\item{L}{Maximum number of pixels an object can move in two consecutive frames}

\item{R}{Linkrange, i.e. the number of consecutive frames to search for potential candidate
links}

\item{epsilon1}{A numeric value, to be used in the formula.
Jitter for allowing angular displacements}

\item{epsilon2}{A numeric value, to be used in the formula.
Jitter for allowing spatial displacements}

\item{lambda1}{A numeric value. Multiplicative factor for the penalty function}

\item{lambda2}{A numeric value. Multiplicative factor applied to the angular displacement}

\item{penaltyFunction}{A function structured in such a way to be applied as penalty function
in the linking}

\item{include.area}{Logical, whether to include also area change of the particles
in the cost function calculation}

\item{frames}{The \code{Frames} object that the \code{ParticleSet} object is
derived from}

\item{verboseOutput}{Logical, whether the output should report additional intermediate steps.
For debugging use mainly.}

\item{prog}{Logical, whether the a progress bar should be shown during the tracking phase}

\item{include.intensity}{Logical, whether to include also intensity change of the particles
in the cost function calculation}
}
\value{
A \code{TrajectorySet} object
}
\description{
Returns a TrajectorySet so that the y value is the distance from the bottom of the
Frames object, not from the top
}
\examples{
library(flowcatchR)
platelets <-particles(channel.Frames(MesenteriumSubset,"red"))
trajSet <- generateTraj(platelets,
                       L=26, R=3,
                       epsilon1=0, epsilon2=0,
                       lambda1=1, lambda2=0,
                       penaltyFunction=penaltyFunctionGenerator(),
                       include.area=FALSE, MesenteriumSubset)
}
\author{
Ayan Bandyopadhyay, Bellarmine College Prep 11/26/2015
}

