% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vidProcessing.R
\name{greenThresh}
\alias{greenThresh}
\title{Thresholds a \code{Frames} object for green pixels}
\usage{
greenThresh(frames)
}
\arguments{
\item{frames}{A \code{Frames} object with all 3 color channels(R,G,B)}
}
\value{
a binary \code{Frames} object in the green color channel
}
\description{
Green pixels, or pixels with H value between 61/360 and 140/360 and
S value over 0.15 in the HSV color scheme, are treated as foreground.
All other pixels are treated as background. This is useful in detecting the
Green Flourescence Protein used to denote that a cell is alive.
}
\examples{
library(flowcatchR)
threshedFrames <- greenThresh(MesenteriumSubset)
}
\author{
Ayan Bandyopadhyay, Bellarmine College Prep 11/26/2015
}

