% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vidProcessing.R
\name{trajMeasures}
\alias{trajMeasures}
\title{Generates 24 measures for each trajectory}
\usage{
trajMeasures(trajSet)
}
\arguments{
\item{trajSet}{This is a \code{TrajectorySet} object,}
\item{spf}{Numeric. Seconds elapsed per frame.}
}
\value{
A matrix 24 columns wide. Each row corresponds to one trajectory.
}
\description{
Removes trajectories with less than 4 data points
}
\examples{
library(flowcatchR)
trajPlatelets <- trajectories(particles(channel.Frames(MesenteriumSubset,"red")))
trajData <-trajMeasures(trajPlatelets, 1)
}
\author{
Ayan Bandyopadhyay, Bellarmine College Prep 11/26/2015
}

